fix: consolidate CLI packages to satisfy pkg-file-count limits
Merge workflow root wiring into root_work.go, combine sessionexecution
inspection commands and loopback tests, and fold human-render helpers into
run.go so pkg/cli and pkg/cli/sessionexecution stay within the 15-file cap.
